refactor: QueryModes & querymodes funtions transfered to xcontext
