+++ internal/table/conn/stmt.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/stack"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/table/conn/badconn"
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xcontext"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xerrors"
 	"github.com/ydb-platform/ydb-go-sdk/v3/trace"
 )
@@ -38,7 +39,7 @@ func (stmt *stmt) QueryContext(ctx context.Context, args []driver.NamedValue) (_
 	if !stmt.conn.isReady() {
 		return nil, badconn.Map(xerrors.WithStackTrace(errNotReadyConn))
 	}
-	switch m := QueryModeFromContext(ctx, stmt.conn.defaultQueryMode); m {
+	switch m := xcontext.QueryModeFromContext(ctx, stmt.conn.defaultQueryMode); m {
 	case DataQueryMode:
 		return stmt.processor.QueryContext(stmt.conn.withKeepInCache(ctx), stmt.query, args)
 	default:
@@ -57,7 +58,7 @@ func (stmt *stmt) ExecContext(ctx context.Context, args []driver.NamedValue) (_ 
 	if !stmt.conn.isReady() {
 		return nil, badconn.Map(xerrors.WithStackTrace(errNotReadyConn))
 	}
-	switch m := QueryModeFromContext(ctx, stmt.conn.defaultQueryMode); m {
+	switch m := xcontext.QueryModeFromContext(ctx, stmt.conn.defaultQueryMode); m {
 	case DataQueryMode:
 		return stmt.processor.ExecContext(stmt.conn.withKeepInCache(ctx), stmt.query, args)
 	default:
